requirements.txt
# Minimum Django and REST framework version
Django>=1.9
djangorestframework>=3.3.2

# Test requirements
pytest-django==2.6
pytest==2.5.2
pytest-cov==1.6
flake8==2.5.4
twotest==0.9.2
